Minor refactor, add various javadocs
