+++ worldedit-core/src/main/java/com/fastasyncworldedit/core/function/mask/AdjacentAny2DMask.java
@@ -8,17 +8,32 @@ import com.sk89q.worldedit.function.mask.Mask;
 import com.sk89q.worldedit.math.BlockVector3;
 
 /**
- * Just an optimized version of the Adjacent Mask for single adjacency.
+ * Optimized version of {@link Adjacent2DMask} for testing for any single adjacency
+ *
+ * @since TODO
  */
 public class AdjacentAny2DMask extends AbstractMask {
 
     private final Mask mask;
     private final MutableBlockVector3 mutable;
 
+    /**
+     * Optimized version of {@link Adjacent2DMask} for testing for any single adjacency. Caches results of the adjacent mask
+     *
+     * @param mask Mask required to be adjacent
+     * @since TODO
+     */
     public AdjacentAny2DMask(Mask mask) {
         this(mask, true);
     }
 
+    /**
+     * Optimized version of {@link Adjacent2DMask} for testing for any single adjacency
+     *
+     * @param mask  Mask required to be adjacent
+     * @param cache If the result of the adjacency mask should be cached
+     * @since TODO
+     */
     public AdjacentAny2DMask(Mask mask, boolean cache) {
         this.mask = cache ? CachedMask.cache(mask) : mask;
         mutable = new MutableBlockVector3();
@@ -41,11 +56,20 @@ public class AdjacentAny2DMask extends AbstractMask {
         return mask.test(mutable.setComponents(x, y, z - 1));
     }
 
-    public boolean test(Extent extent, BlockVector3 v) {
-        AbstractExtentMask extentMask = (AbstractExtentMask) mask;
-        int x = v.x();
-        int y = v.y();
-        int z = v.z();
+    /**
+     * Test this mask for the given extent
+     *
+     * @param extent   extent to test in
+     * @param position position to test at
+     * @since TODO
+     */
+    public boolean test(Extent extent, BlockVector3 position) {
+        if (!(mask instanceof AbstractExtentMask extentMask)) {
+            throw new UnsupportedOperationException("Adjacency mask must inherit from AbstractExtentMask");
+        }
+        int x = position.x();
+        int y = position.y();
+        int z = position.z();
         if (extentMask.test(extent, mutable.setComponents(x + 1, y, z))) {
             return true;
         }
