+++ worldedit-core/src/main/java/com/sk89q/worldedit/EditSessionBuilder.java
@@ -421,6 +421,8 @@ public final class EditSessionBuilder {
 
     /**
      * Set the side effects to be used with this edit
+     *
+     * @since TODO
      */
     public EditSessionBuilder setSideEffectSet(@Nullable SideEffectSet sideEffectSet) {
         this.sideEffectSet = sideEffectSet;
@@ -752,6 +754,8 @@ public final class EditSessionBuilder {
 
     /**
      * Get the SideEffectSet that will be used
+     *
+     * @since TODO
      */
     public SideEffectSet getSideEffectSet() {
         return sideEffectSet;
